Update patch set 1
Patch Set 1:

(3 comments)

We have a design issue with kafka-broker when we want to use it as a storage for web sessions. If one of the nodes is restarted it needs to load sessions from the storage. 
Default implementation just don't have this functionality, websession-flatfile is reading from the file system. 
Websession-broker should get it from the broker implementation(currently from kafka). With kafka-broker in theory it should be easy, we have sessions in the topic we just need to replay them. But because we are using events-broker api it's getting a bit more complicated.
Currently there is no functionality to reset offset for group id on a topic. Maybe we should add to  BrokerApi function like: redeliverAllMessages(topicName)?

Another option is to change current websession-broker code to be just a wrapper for default cache or cache coming from websession-flatfile. So every node is storing local copy(by default in memory, when websession-flatfile provided in file system) after restart local cache is populated without events-broker. But this is more complicated solution and it requires changes in websession-flatfile and probably in gerrit core as well. 

I prefer the first solution as simpler to implement and to setup. WDYT?

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@d5d70762-12d0-45a1-890d-524b12d3f735>
Reviewer: Gerrit User 1012541 <1012541@d5d70762-12d0-45a1-890d-524b12d3f735>
Reviewer: Gerrit User 1015244 <1015244@d5d70762-12d0-45a1-890d-524b12d3f735>
Work-in-progress: false